rename + make damping 0d compatible
